Reinitialized project with correct merged files for sprint 5 v2.3
