--- .bolt/config.json
@@ -1,3 +1,0 @@
-{
-  "template": "bolt-vite-react-ts"
-}
