--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,76 +1,0 @@
-# PR Title: (type): (description) (requirement)  
-> Example: `feat: Add sensor calibration (fr01)`
-
-For checklist, put an `x` in all boxes that apply. the boxes that apply should look like - [x] with no spaces.  
-For other sections, delete the placeholder text (_the italicized lines_) and replace it with your actual content.  
-
-## Type:  
-- [ ] `feat` (frXX / nfrXX) — New feature  
-- [ ] `fix` (bugfix) — Bug fix  
-- [ ] `docs` — Documentation only changes  
-- [ ] `refactor` — Code restructuring without functional changes  
-- [ ] `test` — Adding or updating tests  
-- [ ] `chore` — Maintenance, build, config, or non-feature tasks
-
-_If your PR includes more than one type, choose all that apply and make sure the **title reflects the main purpose**._
-for example: if your PR include both adding new feature and bugfixes: check both and use feat as the title, include both the feature and bug fixes in the description
-
-## Requirement ID:  
-- Single: `(fr01)`  
-- Multiple: `(multi)`  
-- Not Applicable: *(omit parentheses)*  
-
----
-
-## 🧾 Description
-
-**What does this PR do?**  
-_Briefly explain the purpose and content of the pull request._
-
-**Related Components**  
-_List any modules, subsystems, or groups that this change touches or depends on._
-
-**Which group is responsible?**  
-_Indicate your group number (e.g., Group 4 - Web UI)._
-
----
-
-## 🔁 Related Tickets / Issues
-
-_Link any FRs, issues, or tasks associated with this PR._
-
----
-
-## ✅ Checklist
-
-- [ ] PR title follows the naming convention
-- [ ] Type and requirement are correctly marked
-- [ ] My code follows the project's style guidelines
-- [ ] Code compiles/runs correctly
-- [ ] Tests are included/updated if needed
-- [ ] Documentation is updated (if applicable)
-
----
-
-## 🧪 Testing Instructions
-
-_How can reviewers test this PR? Provide steps, screenshots, or test cases._
-
----
-
-## 🔗 Dependencies
-
-_List any external dependencies, such as:_
-_- Related or prerequisite PRs (`#123`, `#456`)_
-_- Specific branches this must be merged after_
-_- Required libraries, modules, or services_
-
-> Example:
-> - Depends on PR #102 (backend API changes)
-> - Uses `v2.0.1` of the temperature-sensor module
-
-- ## 📌 Additional Notes
-
-_Anything reviewers should know, consider, or keep in mind?_
-
----
