--- README.md
@@ -1,2 +1,0 @@
-# Monitoring-Dashboard
-Build HTML/CSS/JS dashboard  for real-time monitoring &amp;  manual control. Modularize  using components.
