+++ frontend/public/resources/user-manual.html
