+++ frontend/src/components/home/VideoTeaser.jsx
