+++ frontend/src/pages/PlantInformation.jsx
