--- src/index.css
@@ -1,76 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  :root {
-    --color-primary: 55 199 121;
-    --color-primary-light: 85 219 151;
-    --color-primary-dark: 45 179 101;
-    
-    --color-secondary: 255 255 255;
-    --color-secondary-light: 255 255 255;
-    --color-secondary-dark: 245 245 245;
-    
-    --color-accent: 25 118 210;
-    --color-accent-light: 66 165 245;
-    --color-accent-dark: 13 71 161;
-    
-    --color-success: 56 142 60;
-    --color-warning: 251 192 45;
-    --color-error: 211 47 47;
-    
-    --color-neutral: 158 158 158;
-    --color-neutral-light: 242 242 242;
-    --color-neutral-dark: 66 66 66;
-  }
-  
-  body {
-    @apply bg-neutral-50 text-neutral-900 font-sans antialiased;
-  }
-  
-  h1, h2, h3, h4, h5, h6 {
-    @apply font-heading font-medium leading-tight;
-  }
-  
-  * {
-    @apply transition-colors duration-200;
-  }
-}
-
-@layer components {
-  .card {
-    @apply bg-white rounded-xl shadow-md p-4 md:p-6 border border-neutral-100;
-  }
-  
-  .btn {
-    @apply px-4 py-2 rounded-lg font-medium transition-all duration-200 
-           focus:outline-none focus:ring-2 focus:ring-offset-2;
-  }
-  
-  .btn-primary {
-    @apply bg-primary hover:bg-primary-dark text-white 
-           focus:ring-primary-light;
-  }
-  
-  .btn-secondary {
-    @apply bg-white hover:bg-neutral-50 text-primary border border-primary
-           focus:ring-primary-light;
-  }
-  
-  .btn-outline {
-    @apply border border-neutral-300 bg-white hover:bg-neutral-50 
-           focus:ring-neutral;
-  }
-  
-  .toggle-switch {
-    @apply relative inline-flex h-6 w-11 items-center rounded-full 
-           transition-colors duration-200 focus:outline-none focus:ring-2 
-           focus:ring-offset-2 focus:ring-primary;
-  }
-  
-  .toggle-switch-circle {
-    @apply inline-block h-4 w-4 transform rounded-full bg-white 
-           transition-transform duration-200;
-  }
-}
