--- src/pages/Signup.jsx
@@ -1,212 +1,0 @@
-import React, { useState, useContext } from 'react';
-import { Link, useNavigate, Navigate } from 'react-router-dom';
-import { Leaf, UserPlus } from 'lucide-react';
-import { AuthContext } from '../App';
-import { auth, db } from '../firebase';  
-import { createUserWithEmailAndPassword, updateProfile, sendEmailVerification } from 'firebase/auth';  
-import { collection, addDoc } from 'firebase/firestore';  
-
-const Signup = () => {
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState(null);  
-  const [successMsg, setSuccessMsg] = useState(null);  
-  const navigate = useNavigate();
-  const { isAuthenticated } = useContext(AuthContext);
-
-  if (isAuthenticated) {
-    return <Navigate to="/dashboard" replace />;
-  }
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setError(null);
-    setSuccessMsg(null);
-
-    try {
-      const userCredential = await createUserWithEmailAndPassword(auth, email, password);
-      const user = userCredential.user;
-
-      await updateProfile(user, { displayName: name });
-
-      await addDoc(collection(db, "User"), {
-        userId: user.uid,
-        name: name,
-        email: email,
-        createdAt: new Date()
-      });
-
-      // Send email verification
-      await sendEmailVerification(user);
-
-      setSuccessMsg("Account created! A verification email has been sent to your email address. Please verify your email before logging in.");
-      
-      // Optionally: navigate to login or dashboard after a delay
-      setTimeout(() => {
-        navigate('/login');
-      }, 5000);
-
-    } catch (err) {
-      console.error("Signup error:", err);
-      setError(err.message);
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-neutral-50 flex flex-col justify-center py-12 sm:px-6 lg:px-8">
-      <div className="sm:mx-auto sm:w-full sm:max-w-md">
-        <div className="flex justify-center">
-          <Link to="/" className="flex items-center gap-2">
-            <div className="relative">
-              <div className="absolute -inset-1 bg-gradient-to-r from-primary to-primary-light rounded-lg blur opacity-25"></div>
-              <div className="relative bg-white rounded-lg p-2">
-                <Leaf size={24} className="text-primary" />
-              </div>
-            </div>
-            <div>
-              <span className="text-xl font-bold bg-gradient-to-r from-primary to-primary-light bg-clip-text text-transparent">
-                Smart
-              </span>
-              <span className="text-xl font-bold text-primary-dark">Grow</span>
-            </div>
-          </Link>
-        </div>
-        <h2 className="mt-6 text-center text-3xl font-bold tracking-tight text-neutral-900">
-          Create your account
-        </h2>
-        <p className="mt-2 text-center text-sm text-neutral-600">
-          Already have an account?{' '}
-          <Link to="/login" className="font-medium text-primary hover:text-primary-dark">
-            Sign in
-          </Link>
-        </p>
-      </div>
-
-      <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
-        <div className="bg-white py-8 px-4 shadow-md sm:rounded-lg sm:px-10">
-          {error && (
-            <div className="mb-4 text-sm text-red-600">
-              {error}
-            </div>
-          )}
-
-          {successMsg && (
-            <div className="mb-4 text-sm text-green-600">
-              {successMsg}
-            </div>
-          )}
-
-          <form className="space-y-6" onSubmit={handleSubmit}>
-            <div>
-              <label htmlFor="name" className="block text-sm font-medium text-neutral-700">
-                Full name
-              </label>
-              <div className="mt-1">
-                <input
-                  id="name"
-                  name="name"
-                  type="text"
-                  autoComplete="name"
-                  required
-                  value={name}
-                  onChange={(e) => setName(e.target.value)}
-                  className="block w-full appearance-none rounded-md border border-neutral-300 px-3 py-2 placeholder-neutral-400 shadow-sm focus:border-primary focus:outline-none focus:ring-primary sm:text-sm"
-                />
-              </div>
-            </div>
-
-            <div>
-              <label htmlFor="email" className="block text-sm font-medium text-neutral-700">
-                Email address
-              </label>
-              <div className="mt-1">
-                <input
-                  id="email"
-                  name="email"
-                  type="email"
-                  autoComplete="email"
-                  required
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  className="block w-full appearance-none rounded-md border border-neutral-300 px-3 py-2 placeholder-neutral-400 shadow-sm focus:border-primary focus:outline-none focus:ring-primary sm:text-sm"
-                />
-              </div>
-            </div>
-
-            <div>
-              <label htmlFor="password" className="block text-sm font-medium text-neutral-700">
-                Password
-              </label>
-              <div className="mt-1">
-                <input
-                  id="password"
-                  name="password"
-                  type="password"
-                  autoComplete="new-password"
-                  required
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  className="block w-full appearance-none rounded-md border border-neutral-300 px-3 py-2 placeholder-neutral-400 shadow-sm focus:border-primary focus:outline-none focus:ring-primary sm:text-sm"
-                />
-              </div>
-            </div>
-
-            <div className="flex items-center">
-              <input
-                id="terms"
-                name="terms"
-                type="checkbox"
-                required
-                className="h-4 w-4 rounded border-neutral-300 text-primary focus:ring-primary"
-              />
-              <label htmlFor="terms" className="ml-2 block text-sm text-neutral-700">
-                I agree to the{' '}
-                <a href="#" className="text-primary hover:text-primary-dark">
-                  Terms of Service
-                </a>{' '}
-                and{' '}
-                <a href="#" className="text-primary hover:text-primary-dark">
-                  Privacy Policy
-                </a>
-              </label>
-            </div>
-
-            <div>
-              <button
-                type="submit"
-                className="flex w-full justify-center items-center gap-2 rounded-md border border-transparent bg-primary py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-primary-dark focus:outline-none focus:ring-2 focus:ring-primary focus:ring-offset-2"
-              >
-                <UserPlus size={18} />
-                Create account
-              </button>
-            </div>
-          </form>
-
-          <div className="mt-6">
-            <div className="relative">
-              <div className="absolute inset-0 flex items-center">
-                <div className="w-full border-t border-neutral-300" />
-              </div>
-              <div className="relative flex justify-center text-sm">
-                <span className="bg-white px-2 text-neutral-500">Or continue with</span>
-              </div>
-            </div>
-
-            <div className="mt-6 grid grid-cols-2 gap-3">
-              <button className="flex w-full items-center justify-center gap-2 rounded-md border border-neutral-300 bg-white py-2 px-4 text-sm font-medium text-neutral-700 shadow-sm hover:bg-neutral-50">
-                Google
-              </button>
-              <button className="flex w-full items-center justify-center gap-2 rounded-md border border-neutral-300 bg-white py-2 px-4 text-sm font-medium text-neutral-700 shadow-sm hover:bg-neutral-50">
-                GitHub
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Signup;
-
