--- tailwind.config.js
@@ -1,53 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
-  theme: {
-    fontFamily: {
-      sans: ['Inter', 'system-ui', '-apple-system', 'BlinkMacSystemFont', 'Segoe UI', 'Roboto', 'Arial', 'sans-serif'],
-      heading: ['Lexend', 'system-ui', '-apple-system', 'BlinkMacSystemFont', 'Segoe UI', 'Roboto', 'Arial', 'sans-serif'],
-    },
-    extend: {
-      colors: {
-        primary: {
-          DEFAULT: 'rgb(var(--color-primary) / <alpha-value>)',
-          light: 'rgb(var(--color-primary-light) / <alpha-value>)',
-          dark: 'rgb(var(--color-primary-dark) / <alpha-value>)',
-        },
-        secondary: {
-          DEFAULT: 'rgb(var(--color-secondary) / <alpha-value>)',
-          light: 'rgb(var(--color-secondary-light) / <alpha-value>)',
-          dark: 'rgb(var(--color-secondary-dark) / <alpha-value>)',
-        },
-        accent: {
-          DEFAULT: 'rgb(var(--color-accent) / <alpha-value>)',
-          light: 'rgb(var(--color-accent-light) / <alpha-value>)',
-          dark: 'rgb(var(--color-accent-dark) / <alpha-value>)',
-        },
-        success: 'rgb(var(--color-success) / <alpha-value>)',
-        warning: 'rgb(var(--color-warning) / <alpha-value>)',
-        error: 'rgb(var(--color-error) / <alpha-value>)',
-        neutral: {
-          DEFAULT: 'rgb(var(--color-neutral) / <alpha-value>)',
-          50: 'rgb(var(--color-neutral-light) / <alpha-value>)',
-          100: '#f5f5f5',
-          200: '#e5e5e5',
-          300: '#d4d4d4',
-          400: '#a3a3a3',
-          500: '#737373',
-          600: '#525252',
-          700: '#404040',
-          800: '#262626',
-          900: '#171717',
-        },
-      },
-      spacing: {
-        '128': '32rem',
-        '144': '36rem',
-      },
-      animation: {
-        'pulse-slow': 'pulse 4s cubic-bezier(0.4, 0, 0.6, 1) infinite',
-      },
-    },
-  },
-  plugins: [],
-};
